requirements.txt
fastapi
Flask
flask-talisman
flask_wtf
uvicorn
gunicorn
requests
pytz
xlrd
openpyxl
sendgrid
boto3
pandas
awscli
pymysql
azure-monitor-query 
azure-identity
python-dotenv